requirements.txt
discord-py-interactions
interactions-persistence
loguru
peewee
dateparser
